+++ CHANGELOG.md
@@ -9,6 +9,10 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Added
+
+- Added `budget` field to experiment spec.
+
 ## [v1.23.0](https://github.com/allenai/beaker-py/releases/tag/v1.23.0) - 2023-12-15
 
 ### Fixed